+++ idea/idea-completion/testData/basic/codeFragments/privatesInSecondPressCompletion.kt
@@ -1,0 +1,13 @@
+import java.io.File
+
+class MyFile : File("file") {
+    private val privateField = 0
+}
+
+fun foo(f: MyFile) {
+    val a = 1<caret>
+}
+
+// INVOCATION_COUNT: 2
+// EXIST: privateField
+// EXIST: prefixLength